♻️ refactor(custom): chatsteps 타입 변경에 따른 폴더 위치 변경
